requirements.txt
pygsheets
requests
pandas
datetime
delorean
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
